Support for multiple sleepers
