+++ kernel/src/interrupts/idt.rs
@@ -21,14 +21,14 @@ use x86_64::{
 use crate::{
     constants::{
         idt::{SYSCALL_HANDLER, TIMER_VECTOR, TLB_SHOOTDOWN_VECTOR},
-        syscalls::{SYSCALL_EXIT, SYSCALL_PRINT},
+        syscalls::{SYSCALL_EXIT, SYSCALL_NANOSLEEP, SYSCALL_PRINT},
     },
-    events::{current_running_event_info, schedule_process, EventInfo},
+    events::inc_runner_clock,
     interrupts::x2apic::{self, current_core_id, TLB_SHOOTDOWN_ADDR},
     memory::{paging::create_mapping, HHDM_OFFSET},
     prelude::*,
-    processes::process::{run_process_ring3, ProcessState, PROCESS_TABLE},
-    syscalls::syscall_handlers::sys_exit,
+    processes::process::try_preempt_process,
+    syscalls::syscall_handlers::{sys_exit, sys_nanosleep},
 };
 
 lazy_static! {
@@ -166,32 +166,54 @@ extern "x86-interrupt" fn page_fault_handler(
 pub extern "x86-interrupt" fn naked_syscall_handler(_: InterruptStackFrame) {
     unsafe {
         naked_asm!(
-            // Push registers to save them
-            "push rax",
-            "push rdi",
-            "push rsi",
-            "push rdx",
-            "push rcx",
-            "push r8",
-            "push r9",
-            "mov	rdi, rsp",
+            // Push registers for potential yielding
+            "
+            push rbp
+            push r15
+            push r14
+            push r13
+            push r12
+            push r11
+            push r10
+            push r9
+            push r8
+            push rdi
+            push rsi
+            push rdx
+            push rcx
+            push rbx
+            push rax
+            ",
+            "mov  rdi, rsp",
             // Call the syscall_handler
             "call syscall_handler",
+
             // Restore registers
-            "pop r9",
-            "pop r8",
-            "pop rcx",
-            "pop rdx",
-            "pop rsi",
-            "pop rdi",
-            "pop rax",
-            "iretq"
+            "
+            pop rax
+            pop rbx
+            pop rcx
+            pop rdx
+            pop rsi
+            pop rdi
+            pop r8
+            pop r9
+            pop r10
+            pop r11
+            pop r12
+            pop r13
+            pop r14
+            pop r15
+            pop rbp
+            iretq
+            "
         );
     }
 }
 
 // This is the actual syscall handler function that reads the registers from the stack
 #[no_mangle]
+#[allow(unused_variables, unused_assignments)]  // disable until args p2-6 are used
 fn syscall_handler(rsp: u64) {
     let syscall_num: u64;
     let p1: u64;
@@ -202,26 +224,19 @@ fn syscall_handler(rsp: u64) {
     let p6: u64;
     let stack_ptr: *const u64 = rsp as *const u64;
     unsafe {
-        syscall_num = *stack_ptr.add(6);
+        syscall_num = *stack_ptr.add(0);
         p1 = *stack_ptr.add(5);
         p2 = *stack_ptr.add(4);
         p3 = *stack_ptr.add(3);
-        p4 = *stack_ptr.add(2);
-        p5 = *stack_ptr.add(1);
-        p6 = *stack_ptr.add(0);
+        p4 = *stack_ptr.add(8);
+        p5 = *stack_ptr.add(6);
+        p6 = *stack_ptr.add(7);
     }
-
-    // temporarily, just print the parameter registers
-    serial_println!("Parameter 1: {}", p1);
-    serial_println!("Parameter 2: {}", p2);
-    serial_println!("Parameter 3: {}", p3);
-    serial_println!("Parameter 4: {}", p4);
-    serial_println!("Parameter 5: {}", p5);
-    serial_println!("Parameter 6: {}", p6);
 
     match syscall_num as u32 {
         SYSCALL_EXIT => sys_exit(),
         SYSCALL_PRINT => serial_println!("Hello world!"),
+        SYSCALL_NANOSLEEP => sys_nanosleep(p1, rsp),
         _ => panic!("Unknown syscall: {}", syscall_num),
     };
 
@@ -278,69 +293,8 @@ extern "x86-interrupt" fn naked_timer_handler(_: InterruptStackFrame) {
 extern "C" fn timer_handler(rsp: u64) {
     inc_runner_clock();
 
-    let event: EventInfo = current_running_event_info();
-    if event.pid == 0 {
-        x2apic::send_eoi();
-        return;
-    }
-
-    // Get PCB from PID
-    let preemption_info = unsafe {
-        let mut process_table = PROCESS_TABLE.write();
-        let process = process_table
-            .get_mut(&event.pid)
-            .expect("Process not found");
-
-        let pcb = process.pcb.get();
-
-        if (*pcb).state != ProcessState::Running || (*pcb).next_preemption_time <= runner_timestamp() {
-            x2apic::send_eoi();
-            return;
-        }
-
-        // save registers to the PCB
-        let stack_ptr: *const u64 = rsp as *const u64;
-
-        (*pcb).registers.rax = *stack_ptr.add(0);
-        (*pcb).registers.rbx = *stack_ptr.add(1);
-        (*pcb).registers.rcx = *stack_ptr.add(2);
-        (*pcb).registers.rdx = *stack_ptr.add(3);
-        (*pcb).registers.rsi = *stack_ptr.add(4);
-        (*pcb).registers.rdi = *stack_ptr.add(5);
-        (*pcb).registers.r8 = *stack_ptr.add(6);
-        (*pcb).registers.r9 = *stack_ptr.add(7);
-        (*pcb).registers.r10 = *stack_ptr.add(8);
-        (*pcb).registers.r11 = *stack_ptr.add(9);
-        (*pcb).registers.r12 = *stack_ptr.add(10);
-        (*pcb).registers.r13 = *stack_ptr.add(11);
-        (*pcb).registers.r14 = *stack_ptr.add(12);
-        (*pcb).registers.r15 = *stack_ptr.add(13);
-        (*pcb).registers.rbp = *stack_ptr.add(14);
-        // saved from interrupt stack frame
-        (*pcb).registers.rsp = *stack_ptr.add(18);
-        (*pcb).registers.rip = *stack_ptr.add(15);
-        (*pcb).registers.rflags = *stack_ptr.add(17);
-
-        (*pcb).state = ProcessState::Blocked;
-
-        ((*pcb).kernel_rsp, (*pcb).kernel_rip)
-    };
-
-    unsafe {
-        schedule_process(event.pid);
-
-        // Restore kernel RSP + PC -> RIP from where it was stored in run/resume process
-        core::arch::asm!(
-            "mov rsp, {0}",
-            "push {1}",
-            in(reg) preemption_info.0,
-            in(reg) preemption_info.1
-        );
-
-        x2apic::send_eoi();
-
-        core::arch::asm!("ret");
-    }
+    try_preempt_process(rsp);
+    x2apic::send_eoi();
 }
 
 // TODO Technically, this design means that when TLB Shootdows happen, each core must sequentially
