+++ kernel/src/processes/test_binaries/sleep.asm
@@ -5,7 +5,7 @@ _start:
     mov rax, 3
     int 0x80
 
-    mov rbx, 5000000000
+    mov rdi, 5000000000
     mov rax, 35
     int 0x80
